Tutorial 4 fix because of new parser brhaviour.
Bit of a hokey patch here, to work around the problem of the template parser no longer allowing access to a parent variable inside a child loop - used for the columns in the menu display. this feature, formerly in CI2.2, was removed in CI3 because it presented a security vulnerability.
Signed-off-by:Master Yoda <jim_parry@bcit.ca>